requirements.txt
alembic==1.9.0
anyio==3.6.2
boto3==1.26.37
botocore==1.29.37
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keyboa==3.1.0
Mako==1.2.4
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
psycopg2==2.9.5
pydantic==1.10.2
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-requests==2.28.11.7
types-urllib3==1.26.25.4
types-Werkzeug==1.0.9
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
WTForms==3.0.1

python-dotenv~=1.0.0
pytest~=7.2.2